backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
python-multipart
passlib
bcrypt
python-dotenv
